+++ src/audioresampler.cpp
@@ -47,7 +47,7 @@ AudioResampler::~AudioResampler() {
 
 std::vector<AVFrame*> AudioResampler::resample_audio_frames(std::vector<AVFrame*>& originals) {
     std::vector<AVFrame*> resampled_frames;
-    for (int i = 0; i < originals.size(); i++) {
+    for (int i = 0; i < (int)originals.size(); i++) {
         AVFrame* resampled_frame = this->resample_audio_frame(originals[i]);
         resampled_frames.push_back(resampled_frame);
     }
